Fix container build
Fix the container build process.

The container build with RUN --mount=type=cache was a good speedup
reducing downloads from internet and more aggressive caching. But it
resulted in unwanted behavior of Docker/OrbStack: A changed mix release from
the build phase was not updated in the final image.

This commit removes the RUN --mount=type=cache for the mix release and
all related RUN commands.

The cache for updating/installing Alpine packages is still in use.